Split HTTP client into Kafka-aligned modules
• Replace client.rs with producer.rs and consumer.rs modules
• Move health commands to metadata.rs module
• Extract error handling to error.rs module
• Move print utilities and tests to common.rs
• Update CLI imports to use new module structure
